refactor : added new section cta
