+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "@headlessui/react": "^2.2.0",
+    "lucide-react": "^0.536.0",
     "next": "^15.4.5",
     "nodemailer": "^7.0.5",
     "react": "^19.0.0",
