+++ src/app/page.tsx
@@ -3,7 +3,7 @@ import CaseStudies from "@/components/CaseStudies";
 import Contact from "@/components/Contact";
 import Footer from "@/components/Footer";
 import Hero from "@/components/Hero";
-import LoginSignup from "@/components/LoginSignup";
+import LoginSignup from "@/components/cta";
 import Services from "@/components/Services";
 import Team from "@/components/Team";
 import Testimonials from "@/components/Testimonials";
